Yet another Windows fix for dependent packages
I've gotten to the point of trying to compile libX11 on Windows, and it errors
out due to undefined names like socklen_t. It turns out that someone needs to
include WS2tcpip.h. The Xwinsock.h header seems like the most reasonable place
to put this #include, and we're patching the file anyway, so let's patch it a
little bit more.